fix(kas): test through extractAllPackageFiles
